training/labelling: Update to handle new depth/label formats

+++ training/labelling/infer.cc
@@ -5,7 +5,7 @@
 #include "loader.h"
 
 float*
-infer(RDTree** forest, uint8_t n_trees, float* depth_image,
+infer(RDTree** forest, uint8_t n_trees, half* depth_image,
       uint32_t width, uint32_t height)
 {
   uint8_t n_labels = forest[0]->header.n_labels;
