+++ training/labelling/infer.h
@@ -3,11 +3,12 @@
 #define __INFER__
 
 #include <stdint.h>
+#include <half.h>
 #include "loader.h"
 
 float* infer(RDTree** forest,
              uint8_t  n_trees,
-             float*   depth_image,
+             half*    depth_image,
              uint32_t width,
              uint32_t height);
